+++ node/index.ts
@@ -1,3 +1,5 @@
+import { Service } from '@vtex/api'
+
 import { queries } from './resolvers/index'
 import { postResolvers } from './resolvers/postResolvers'
 import { pageResolvers } from './resolvers/pageResolvers'
@@ -5,7 +7,6 @@ import { titleResolvers } from './resolvers/titleResolvers'
 import { excerptResolvers } from './resolvers/excerptResolvers'
 import { categoryResolvers } from './resolvers/categoryResolvers'
 import { tagResolvers } from './resolvers/tagResolvers'
-import { Service } from '@vtex/api'
 import { clients } from './clients'
 
 export default new Service({