+++ react/components/WordpressPage.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/camelcase */
 import { Container } from 'vtex.store-components'
-
 import React, { FunctionComponent, useMemo } from 'react'
 import { Helmet } from 'react-helmet'
 import { defineMessages } from 'react-intl'
@@ -230,7 +229,7 @@ const WordpressPage: StorefrontFunctionComponent<PageProps> = ({
 }) => {
   const {
     route: { params },
-  } = useRuntime()
+  } = useRuntime() as any
 
   let parsedCustomDomains = null
   try {